AddRecipe Class:
Changed the StartAddRecipe method signature to accept a list of AddRecipe objects as a parameter instead of being static and not taking any parameters.
Modified the StartAddRecipe method to prompt the user for the number of recipes they want to add before starting the process of adding each recipe.
Changed the type of the recipes variable in the StartAddRecipe method from an array to a list to enable dynamic resizing and better flexibility.
Adjusted the loop in the StartAddRecipe method to iterate over the number of recipes specified by the user and add each recipe to the recipes list.
Modified the StartAddRecipe method to take the recipes list as a parameter, and after adding recipes, it invokes the StartMessage method from the Start class, passing the recipes list as an argument to continue the program flow.
ScaleRecipe Class:

Changed the ScaleQuantities method signature to accept an integer parameter numIngredients.
Removed the dependency on the AddRecipe class properties and fields. Instead of directly accessing properties of the AddRecipe class, the ScaleQuantities method now takes the required ingredients as method arguments.
Removed the dependency on the DisplayRecipe class. The method no longer directly invokes the Display method.
RunPoe Class:

Removed the Start.StartMessage() method call from the Main method. This call is now made from within the StartAddRecipe method of the AddRecipe class.
DisplayRecipe Class:

Added a new method DisplaySpecificRecipe to display a specific recipe based on user input. This method takes a list of AddRecipe objects as a parameter.
Modified the existing Display method to sort recipes by name and display all recipes. This method also takes a list of AddRecipe objects as a parameter.
ClearRecipe Class:

Changed the method name from ClearArrays to ClearRecipes.
Modified the ClearRecipes method to accept a list of AddRecipe objects as a parameter.
Removed the direct clearing of properties and fields of the AddRecipe class. Instead, the method now clears the list of recipes.